requirements.txt
streamlit
openai
python-dotenv
pandas
openpyxl
pillow
pdf2image
